Split AccountsHelper from StatusesHelper
